remove AppConfigurationContext file
